+++ samples/react/src/containers/ActiveConversation.js
@@ -392,11 +392,22 @@ export default class ActiveConversation extends Component {
           });
           model.generateMessage(conversation, message => message.send());
         }.bind(this),
+      },
+      {
+        text: 'Create Feedback Message',
+        method: function() {
+          const FeedbackModel = Layer.Core.Client.getMessageTypeModelClass('FeedbackModel');
+          const model = new FeedbackModel({
+            enabledFor: [client.user.id],
+          });
+          model.generateMessage(conversation, message => message.send());
+        }.bind(this),
       }
     ];
   };
 
   filterMessages(message) {
+    return true;// remove this to enable filtering
     const model = layerClient.createMessageTypeModel(message);
     if (model && model instanceof ResponseModel && message.sender === layerClient.user) return false;
     return true;