+++ src/ui/messages/feedback/test.js
@@ -1,0 +1,72 @@
+describe('Feedback Message Components', function() {
+  var FeedbackModel;
+  var conversation;
+  var testRoot;
+
+  var styleNode;
+  beforeAll(function() {
+    styleNode = document.createElement('style');
+    styleNode.innerHTML = 'layer-message-viewer.layer-feedback-view  {width: 300px; height: 150px;}';
+    document.body.appendChild(styleNode);
+  });
+
+  afterAll(function() {
+    document.body.removeChild(styleNode);
+  });
+
+  beforeEach(function() {
+    jasmine.clock().install();
+    restoreAnimatedScrollTo = layer.UI.animatedScrollTo;
+    spyOn(layer.UI, "animatedScrollTo").and.callFake(function(node, position, duration, callback) {
+      var timeoutId = setTimeout(function() {
+        node.scrollTop = position;
+        if (callback) callback();
+      }, duration);
+      return function() {
+        clearTimeout(timeoutId);
+      };
+    });
+
+    client = new layer.Core.Client({
+      appId: 'layer:///apps/staging/Fred'
+    });
+    client.user = new layer.Core.Identity({
+      client: client,
+      userId: 'FrodoTheDodo',
+      displayName: 'Frodo the Dodo',
+      id: 'layer:///identities/FrodoTheDodo',
+      isFullIdentity: true,
+      sessionOwner: true
+    });
+    client._clientAuthenticated();
+    conversation = client.createConversation({
+      participants: ['layer:///identities/FrodoTheDodo', 'layer:///identities/SaurumanTheMildlyAged']
+    });
+
+    if (layer.UI.components['layer-conversation-view'] && !layer.UI.components['layer-conversation-view'].classDef) layer.UI.init({});
+
+    testRoot = document.createElement('div');
+    document.body.appendChild(testRoot);
+    testRoot.style.display = 'flex';
+    testRoot.style.flexDirection = 'column';
+    testRoot.style.height = '300px';
+
+    FeedbackModel = layer.Core.Client.getMessageTypeModelClass("FeedbackModel");
+
+    layer.Util.defer.flush();
+    jasmine.clock().tick(800);
+    jasmine.clock().uninstall();
+  });
+
+
+  afterEach(function() {
+    layer.UI.animatedScrollTo = restoreAnimatedScrollTo;
+    layer.Core.Client.removeListenerForNewClient();
+  });
+
+  describe("Model Tests", function() {
+    it("Should create an appropriate Basic Model with suitable properties and defaults", function() {
+      expect("Should have tests").toEqual("Yeah right...");
+    });
+  });
+});